Avoid closing socket for each request
This commit adds an internal Hashtable to cache the RadiusClient object
for each thread. The main reason is to reuse the socket. With the old
implementation a socket was opened for every single request which
quickly leads to socket exhaustion on the JMeter machine when trying
to reach a reasonable throughput.